requirements.txt
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pandas>=2.1.0
streamlit>=1.29.0
python-dotenv>=1.0.0
httpx>=0.25.0
requests>=2.31.0
plotly>=5.18.0
tqdm>=4.66.0
pytz>=2023.3
databento>=0.36.0
vaderSentiment>=3.3.2
statsmodels>=0.14.6
